git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@3989 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/geometry/Sphere.hpp
@@ -57,7 +57,7 @@ namespace geometry {
 
 	private:
 		double _radius;
-		int _levels;
+		double _levels;
 	};
 	//! @}
